requirements.txt
setuptools
discord.py
pyautogui
